# api/openapi.yaml
openapi: 3.0.3
info:
  title: FQP From Scratch API
  version: 1.0.0
  description: 官方足彩/竞彩足球量化研究与长期运行系统 API 草案。系统只提供数据分析、模拟与复盘接口，不提供购彩、代购、出票、收款接口。
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/models/predict/date/{date}:
    post:
      summary: 生成指定日期模型预测
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        '202': { description: 预测任务已提交 }
  /api/v1/models/predictions:
    get:
      summary: 查询模型预测结果
      parameters:
        - name: match_id
          in: query
          required: false
          schema: { type: integer }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: 模型预测列表 }
...
  /api/v1/predictions:
    get:
      summary: 查询当前实现的模型预测列表
      parameters:
        - name: match_id
          in: query
          required: false
          schema: { type: integer }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: 模型预测列表 }
  /api/v1/recommendations/generate:
    post:
      summary: 生成每日推荐模拟票单
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: 推荐结果或熔断原因 }
  /api/v1/recommendations/daily:
    get:
      summary: 查询每日推荐
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: 每日推荐票单 }
...
  /api/v1/recommendations/{ticket_id}/invalidate:
    post:
      summary: 作废推荐票单
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 已作废或返回不可作废原因 }
  /api/v1/tickets:
    get:
      summary: 查询模拟推荐票单
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200': { description: 模拟推荐票单列表 }
...